+++ packages/crud-core/package.json
@@ -13,9 +13,8 @@
     "./shared/crudFieldMetaSupport": "./src/shared/crudFieldMetaSupport.js",
     "./shared/crudNamespaceSupport": "./src/shared/crudNamespaceSupport.js",
     "./server/repositorySupport": "./src/server/repositorySupport.js",
-    "./server/repositoryMethods": "./src/server/repositoryMethods.js",
+    "./server/repositoryOrm": "./src/server/repositoryOrm/index.js",
     "./server/createHooksToCollectChildren": "./src/server/createHooksToCollectChildren.js",
-    "./server/createCrudRepositoryFromResource": "./src/server/createCrudRepositoryFromResource.js",
     "./server/lookupProviders": "./src/server/lookupProviders.js",
     "./server/serviceEvents": "./src/server/serviceEvents.js",
     "./server/serviceMethods": "./src/server/serviceMethods.js",
