+++ packages/users-core/test/usersRouteResources.test.js
@@ -11,7 +11,7 @@ import { userSettingsResource } from "../src/shared/resources/userSettingsResour
 function assertResourceShape(resource, label) {
   assert.ok(resource, `${label} resource must exist.`);
   assert.equal(typeof resource, "object", `${label} resource must be an object.`);
-  assert.equal(typeof resource.resource, "string", `${label}.resource must be a string.`);
+  assert.equal(typeof resource.namespace, "string", `.namespace must be a string.`);
 
   for (const operationName of ["view", "list", "create", "replace", "patch"]) {
     const operation = resource.operations?.[operationName];
