+++ tooling/jskit-catalog/catalog/packages.json
@@ -1838,7 +1838,7 @@
             "inputType": "text",
             "defaultValue": "",
             "promptLabel": "Namespace override",
-            "promptHint": "Optional CRUD namespace override when the resource export does not expose resource.resource."
+            "promptHint": "Optional CRUD namespace override when the resource export does not expose resource.namespace."
           }
         },
         "dependsOn": [],
